Reorg the code to move generic stuff out of ferc1.py
