+++ src/pudl/ferc_to_sqlite/__init__.py
@@ -4,7 +4,7 @@ import importlib.resources
 from dagster import Definitions, graph
 
 import pudl
-from pudl.extract.ferc1 import dbf2sqlite
+from pudl.extract.ferc import dbf2sqlite
 from pudl.extract.xbrl import xbrl2sqlite
 from pudl.resources import datastore, ferc_to_sqlite_settings
 from pudl.settings import EtlSettings
